refactor: arquivo separado para factories
