+++ tests/conftest.py
@@ -1,37 +1,26 @@
-import factory
 import pytest
 from fastapi.testclient import TestClient
-from sqlalchemy import StaticPool, create_engine
+from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
 from fast_todo.app import app
 from fast_todo.database import get_session
-from fast_todo.models import Base, User
-from fast_todo.settings import Settings
+from fast_todo.models import Base
 from fast_todo.security import get_password_hash
-
-
-class UserFactory(factory.Factory):
-    class Meta:
-        model = User
-
-    id = factory.Sequence(lambda n: n)
-    username = factory.LazyAttribute(lambda obj: f'test{obj.id}')
-    email = factory.LazyAttribute(lambda obj: f'{obj.username}@test.com')
-    password = factory.LazyAttribute(lambda obj: f'{obj.username}@example.com')
+from fast_todo.settings import Settings
+from tests.factories import UserFactory
 
 
 @pytest.fixture
 def session():
-    engine = create_engine(
-        Settings().DATABASE_URL,
-        poolclass=StaticPool,
-    )
+    engine = create_engine(Settings().DATABASE_URL)
 
-    Session = sessionmaker(bind=engine)
+    Session = sessionmaker(autocommit=False, autoflush=False, bind=engine)
     Base.metadata.create_all(engine)
 
-    yield Session()
+    with Session() as session:
+        yield session
+        session.rollback()
 
     Base.metadata.drop_all(engine)
